requirements.txt
google-cloud-aiplatform>=1.38.0
google-cloud-storage>=2.14.0
google-cloud-datacatalog>=3.16.0
pydantic>=2.0.0
python-dotenv>=1.0.0
